chore: color token 상수화 및 theme에 colors 추가
