+++ zoo-client/src/styles/tokens/index.ts
@@ -1,0 +1,1 @@
+export { default as colors } from './colors';
